tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-location-replace.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="require-trusted-types">
</head>
<body>
<script>
  //TrustedURL replacements work
  test(t => {
    var url = TrustedURL.create(location.href + "#xxx");
    location.replace(url);
    assert_equals("" + url, location.href, "location href");
  }, "Basic processing: safe URL, safe construction.");

  test(t => {
    var url = TrustedURL.unsafelyCreate(location.href + "#xxx");
    location.replace(url);
    assert_equals("" + url, location.href, "location href");
  }, "Basic processing: safe URL, unsafe construction.");

  // String replacements throw.
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace("A string");
    });
  }, "`location.replace = string` throws");

  //Null replacement throws
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace(null);
    });
  }, "`location.replace = null` throws");
</script>
